Add fit reader to calculate the midpoint of a route

+++ src/test/java/com/github/dballinger/activityweather/DecimalDegreeTest.java
@@ -10,7 +10,7 @@ public class DecimalDegreeTest {
     @Test
     public void shouldConvertDecimalDegreeToSemicircle() throws Exception {
         double degrees = 51.238820096477866;
-        long semicircles = 611302935;
+        int semicircles = 611302935;
         DecimalDegree decimalDegree = new DecimalDegree(degrees);
         assertThat(decimalDegree.toSemicircle(), is(new Semicircle(semicircles)));
     }
