+++ src/test/java/com/github/dballinger/activityweather/SemicircleTest.java
@@ -10,7 +10,7 @@ public class SemicircleTest {
     @Test
     public void shouldConvertSemicircleToDecimalDegree() throws Exception {
         double degrees = 51.238820096477866;
-        long semicircles = 611302935;
+        int semicircles = 611302935;
         Semicircle semicircle = new Semicircle(semicircles);
         assertThat(semicircle.toDegrees(), is(new DecimalDegree(degrees)));
     }
